Refine constants in docstring
