+++ docs/how-to-guides/develop-a-flow/index.md
@@ -3,7 +3,6 @@ We provide guides on how to develop a flow by writing a flow yaml from scratch i
 
 ```{toctree}
 :maxdepth: 1
-:hidden:
 
 init-and-test-a-flow
 develop-standard-flow
