+++ scripts/docs/api_doc_templates/package.rst_t
@@ -16,7 +16,7 @@
 {%- if is_namespace %}
 {{- [pkgname, "namespace"] | join(" ") | e | heading }}
 {% else %}
-{{- [pkgname, ""] | join(" ") | e | heading }}
+{{- [pkgname, "module"] | join(" ") | e | heading }}
 {% endif %}
 
 {%- if is_namespace %}
